LayoutTests/platform/ios-simulator/fast/canvas/canvas-crash-expected.txt

CONSOLE MESSAGE: line 16: Canvas area exceeds the maximum limit (width * height > 16777216).
PASSED 
